fix : PercentDiscount 계산 수정

+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/model/FixedAmountVoucher.java
@@ -14,7 +14,7 @@ public class FixedAmountVoucher implements Voucher {
     return voucherId;
   }
 
-  public long discountedPrice(long price) {
+  public double discountedPrice(long price) {
     return price - discount;
   }
