+++ springboot-basic/src/main/java/com/example/commandlineapplication/domain/voucher/model/PercentDiscountVoucher.java
@@ -6,17 +6,17 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class PercentDiscountVoucher implements Voucher {
 
-  private static final int FULL_PERCENT = 100;
+  private static final double FULL_PERCENT = 100;
   private final UUID voucherId;
   private final long discountPercent;
-  
+
   @Override
   public UUID getVoucherId() {
     return voucherId;
   }
 
   @Override
-  public long discountedPrice(long price) {
+  public double discountedPrice(long price) {
     return price * (discountPercent / FULL_PERCENT);
   }
